login

+++ controllers/usuario_controllers.py
@@ -1,17 +1,23 @@
-from flask import request, redirect, url_for, Blueprint
+from flask import request, redirect, url_for, Blueprint, session, abort
+from utils.decorators import login_required, role_required
 
 from models.usuario_model import Usuario
 from views import usuario_view
 
-usuario_bp = Blueprint('usuario',__name__,url_prefix="/usuarios")
+usuario_bp = Blueprint('usuario', __name__, url_prefix="/usuarios")
+
 
 @usuario_bp.route("/")
+@login_required
+@role_required('admin', 'bibliotecario')
 def index():
-    #recupera todos los registros de la tabla usuario en forma de objeto
     usuarios = Usuario.get_all()
     return usuario_view.list(usuarios)
 
-@usuario_bp.route("/create", methods = ['GET','POST'])
+
+@usuario_bp.route("/create", methods=['GET', 'POST'])
+@login_required
+@role_required('admin')  # Solo admin puede crear
 def create():
     if request.method == 'POST':
         nombre = request.form['nombre']
@@ -27,9 +33,17 @@ def create():
         return redirect(url_for('usuario.index'))
     return usuario_view.create()
 
-@usuario_bp.route("/edit/<int:id>", methods = ['GET','POST'])
+
+@usuario_bp.route("/edit/<int:id>", methods=['GET', 'POST'])
+@login_required
+@role_required('admin', 'bibliotecario')
 def edit(id):
     usuario = Usuario.get_by_id(id)
+
+    # Solo admin puede editar todo; bibliotecario solo puede editar lectores
+    if session['tipo'] == 'bibliotecario' and usuario.tipo != 'lector':
+        abort(403)  # Prohibido
+
     if request.method == 'POST':
         nombre = request.form['nombre']
         apellido = request.form['apellido']
@@ -38,14 +52,27 @@ def edit(id):
         username = request.form['username']
         password = request.form['password']
         tipo = request.form['tipo']
-        #actualizar
-        usuario.update(nombre=nombre, apellido=apellido, email=email, telefono=telefono, username=username, password=password, tipo=tipo)
+
+        # Verificación para que bibliotecario no pueda cambiar el tipo a admin o bibliotecario
+        if session['tipo'] == 'bibliotecario' and tipo != 'lector':
+            abort(403)
+
+        usuario.update(nombre=nombre, apellido=apellido, email=email,
+                       telefono=telefono, username=username, password=password, tipo=tipo)
         return redirect(url_for('usuario.index'))
-    
+
     return usuario_view.edit(usuario)
 
+
 @usuario_bp.route("/delete/<int:id>")
+@login_required
+@role_required('admin', 'bibliotecario')
 def delete(id):
     usuario = Usuario.get_by_id(id)
+
+    # Solo admin puede eliminar admin o bibliotecario
+    if session['tipo'] == 'bibliotecario' and usuario.tipo != 'lector':
+        abort(403)
+
     usuario.delete()
     return redirect(url_for('usuario.index'))
